Pinentry timeout support.
This adds a new pinentry command SETTIMEOUT and pinentry command line
option --timeout to set the number of seconds until the pinentry dialog
closes itself (default is 60 seconds). Only the qt4 program has been
modified to make use of it.

The timer is cancelled when user input is detected. For confirmation
dialog boxes with a Cancel button, the button press is emulated after the
timer expires. For a confirmation box without a Cancel button, nothing
happens.