w32: Minor changes to the dialog.
* w32/main.c (set_bitmap): New.
(dlg_proc): Show error prompt in red.
* w32/pinentry-w32.rc: Add icons and chnage dialog.
* w32/logo-128.bmp: New.
* w32/logo-32.bmp: New.
* w32/logo-48.bmp: New.
* w32/logo-64.bmp: New.
* w32/logo-96.bmp: New.

--

The logos need some manual tweaking.  Due to the scaling a gradient is
used for the circle and Windows shows white ugly white spots.  This
can be avoided by using a a palette with less colors but I found no
easy way to do that in gimp.  When using less than 256 index the
transparency feature does not work correctly.