Add option "invisible-char".
* pinentry/pinentry.h (struct pinentry): Add field invisible_char.
* pinentry/pinentry.c (pinentry_reset): Restore that.
(option_handler): Add option "invisible-char".
* gtk+-2/pinentry-gtk-2.c (create_window): Set the invisible char if
given.
--

The default GTK+ invisible character is pretty wide so that only ~16
characters show up in the smalles pinentry.  This is a bit annoying if
the passphrase is a longer due to the missing visual feedback.  By
using #x2022 ("•") this allows for ~23 characters and a plain "*" even
3 more characters.  Tastes are different so we allow to change that
character at runtime.

Signed-off-by: Werner Koch <wk@gnupg.org>