Qt: Add SetWindowPos based foreground hack for Win
* qt/pinentrydialog.cpp(raiseWindow): Add another fallback for
our foreground window hacks.

--
Even if SetForegroundWindow or SetForegroundWindowEx do not report
failures we are not always brought to front. So additionally
afterwards we also set our Window Position to be absolutely
in foreground and afterards remove that (so that a user
may still but us in the background).

This fixes the weird behavior that repeated pinentries for
symmetric encryption open in background.