gtk: Really always set the window as transient.
* gtk+-2/pinentry-gtk-2.c (make_transient): Set the window as
transient even if we do not grab the keyboard.
--

In the previous commit attempting to fix issue 3253, the window
was set as transient only in grabbing mode. The window should
actually always be set as transient, independently of whether
we grab the keyboard or not. This is especially important now
that --no-grab is the default behavior of GnuPG Agent.

GnuPG-bug-id: 3253
Fixes-commit: f69dadc6ccea7672869436291ab5c1f58d545466
Signed-off-by: Damien Goutte-Gattat <dgouttegattat@incenp.org>