gnome3: Use the default dbus timeout when checking for screenlock
* gnome3/pinentry-gnome3.c (pe_gnome_screen_locked): use default dbus
timeout instead of 0ms.

--

At the moment, every single check to see if the screen is locked fails
immediately with a timeout, which is then silently ignored.

This is happening because the definition of the timeout argument is a
millisecond count, with -1 as the (unspecified) default, and G_MAXINT for no
timeout.

This means 0 times out immediately, and that leads to a complete inability for
pinentry-gnome3 to fallback to curses input when the screen is locked.

(Which then prevents a remote session from ever asking for a PIN, even when
the local session is locked.)

Debian-bug-id: 927105
Signed-off-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>