qt: Improve focus and frontness
* qt/pinentrydialog.cpp (raiseWindow): Use QtWindowsWindowFunctions
to get a more agressive bring to frontness.
(PinEntryDialog::PinEntryDialog): Start minimized.

--
Weirdly enough this works much better then the old code because
we now start minimized and then bring us to front when the
event loop is fully listening. This way we have the
allowsetforegroundwindow right because we are restored from
minimized state. This helps even in cases where
the QtWindowsWindow Behaviror did not activate the window
besides the requested behavior change.

GnuPG-Bug-Id: T4123