qt: Show hint if passphrase is shown and formatting is enabled
* qt/pinentrydialog.cpp (PinEntryDialog::PinEntryDialog): Initialize
new fields.
(PinEntryDialog::PinEntryDialog): Add new widgets to GUI.
(PinEntryDialog::setFormattedPassphrase): Set text of the hint.
(PinEntryDialog::toggleFormattedPassphrase): Show/hide hint and spacer
as needed.
* qt/pinentrydialog.h (PinEntryDialog): Add fields
mFormattedPassphraseHint, mFormattedPassphraseHintSpacer.
--

The spacer is used to avoid unnecessary layout changes. It ensures that
the vertical space taken by the hint is retained when the hint is hidden.

GnuPG-bug-id: 5517