qt: Embed icons to avoid theme searches
* m4/qt.m4: Check for RCC tool
* qt/qrc_pinentry.cpp: Remove not used file.
* qt/document-encrypt.png: Move to icons subfolder
* qt/icons/*.svg: New.
* qt/main.cpp: Use new path for window icon.
* qt/pinentryrc.qrc: Real ressource file.
* qt/pinentrydialog.cpp: Use embedded icons.

--
This adds the icons from the breeze icon set under a compatible
license. These icons were already used through the icon theme
engine. Embedding them speeds up the startup of pinentry-qt on
my system bei 160ms as the whole Theme load is skipped. This
can be noticable and having pinentry-qt search lots of disk
paths for theme icons was wasteful.

The old qrc_pinentry.cpp looked like a try in the past to do
this but was not used currently.