qt: Create all widgets in tab order
* qt/pinentrydialog.cpp (PinEntryDialog::PinEntryDialog): Initialize
_have_quality_bar with member initializer; rework setup of UI
--

Create all widgets in the order in which they are added to the layout.
This ensures that the tab order matches the visual order of the widgets.
This change also adds the widgets to the layout immediately after
creation instead of first creating all widgets and then adding them
to the layout. And it moves all connects after the creation of the
widgets.

Note: Before this commit the tab order of the usual focus widgets (input
fields, buttons, etc.) was already correct. Just a few labels were
created out-of-order.

GnuPG-bug-id: 5863