qt: Use better icon for signalling an error
* qt/icons/data-error.svg: New.
* qt/icons/Makefile.am, qt/pinentryrc.qrc: Add new file.
* qt/pinentrydialog.h, qt/pinentrydialog.cpp (icon): Rename to
applicationIconPixmap. Change optional parameter to QIcon.
* qt/pinentrydialog.cpp (PinEntryDialog::PinEntryDialog,
PinEntryDialog::setDescription): Adjust to renaming of icon().
(PinEntryDialog::setError): Use new icon as overlay for application icon.
* qt/main.cpp (qt_cmd_handler):  Adjust to renaming of icon().
--

This replaces the style's QStyle::SP_MessageBoxCritical icon of the used
icon theme with the data-error.svg of the Breeze icon theme (originally
licensed under the LGPL v3 or later) as overlay for the application icon
if an error message is shown.

GnuPG-bug-id: 5867