qt: Fix reaction to Escape key and close button of window
* qt/pinentryconfirm.cpp (PinentryConfirm::showEvent): Call the
overridden method of the direct super class.
--

This makes the Escape key and the close button of the window work as
expected resp. work at all. QMessageBox auto-detects the button to press
if the Escape key is pressed or the window close button is clicked in
QMessageBox:;showEvent(). It also does other things in this method, so
we better let it do its magic.

GnuPG-bug-id: 5863