qt: Add dark mode icon variants and detection
* qt/icons/hint_dark.svg, qt/icons/password-generate_dark.svg,
qt/icons/visibility_dark.svg: New.
* qt/icons/Makefile.am (EXTRA_DIST): Add icons.
* qt/pinentrydialog.cpp (PinEntryDialog::PinEntryDialog):
Check if colorScheme is dark. Use Icons accordingly.
* qt/pinentrydialog.h (PinentryDialog): Add icon suffix member
variable.
* qt/pinentryrc.qrc: Add new icons.

--
This uses Qt 6.5 QGuiApplication::styleHints()::colorScheme()
to check for dark mode and suffixes the icon names with _dark
to load their counterparts if this is required.

Tested on Windows 10 2016 and 22h2 as well as plasma / KDE.

GnuPG-Bug-Id: T7230